fix physics-graphics syncronization
the `drawrect` function draws the paddle at x=0 if x<0; this leads to the paddle getting out of the window physically, whereas the graphical window shows it inside the window. so that if you then press the right arrow, you get a delay in paddle movement ( you wait till x > 0). If we don't allow the physical paddle to get out of the window, problem solved!